requirements.txt
torch>=2.0.0
transformers>=4.51.0
requests>=2.28.0
python-dotenv>=1.0.0
llama-cpp-python>=0.3.0
